move delimiter back into input package
