+++ input/collector.go
@@ -10,8 +10,6 @@ import (
 	"strings"
 	"sync/atomic"
 	"time"
-
-	"github.com/ekanite/ekanite/input/delimiter"
 )
 
 var sequenceNumber int64
@@ -115,7 +113,7 @@ func (s *TCPCollector) handleConnection(conn net.Conn, c chan<- *Event) {
 		conn.Close()
 	}()
 
-	delimiter := delimiter.NewSyslogDelimiter(msgBufSize)
+	delimiter := NewSyslogDelimiter(msgBufSize)
 	reader := bufio.NewReader(conn)
 	var log string
 	var match bool
