+++ input/syslog_delimiter.go
@@ -1,4 +1,4 @@
-package delimiter
+package input
 
 import (
 	"regexp"
