+++ input/syslog_delimiter_test.go
@@ -1,4 +1,4 @@
-package delimiter
+package input
 
 import (
 	"testing"
